requirements.txt
cycler==0.10.0
Django==1.11.4
django-admin-view-permission==1.1
django-ckeditor==5.3.0
django-js-asset==0.1.1
django-reversion==2.0.10
matplotlib==2.0.2
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
setuptools-git==1.2
six==1.10.0
